Add tokens to identify legit lock owner
Previously, everyone could unlock a machine locked by someone else.
Unfortunately, when we unlock the machine in a post-fail-hook in openQA,
we will unlock any machine, even if it is locked, and even if we just
failed to lock it.
This commit introduces randomly generated tokens, which need to be
passed when unlocking a host, otherwise the unlock will fail.

Signed-off-by: Michael Moese <mmoese@suse.de>